feat: refactored back the paymaster to a hook
