+++ frontend/src/components/modals/BuySeeds.tsx
@@ -19,6 +19,7 @@ import {
   useNetwork,
   useBalance,
 } from "@fuels/react";
+import { usePaymaster } from "../../hooks/usePaymaster";
 
 interface BuySeedsProps {
   contract: FarmContract | null;
@@ -43,42 +44,11 @@ export default function BuySeeds({
       throw new Error("No wallet found");
     }
     const provider = await Provider.create(FUEL_PROVIDER_URL);
-    const { data: MetaDataResponse } = await axios.get<{
-      maxValuePerCoin: string;
-    }>(`http://167.71.42.88:3000/metadata`);
-    const { maxValuePerCoin } = MetaDataResponse;
-    console.log("maxValuePerCoin", maxValuePerCoin);
-    if (!maxValuePerCoin) {
-      throw new Error("No maxValuePerCoin found");
-    }
-    const { data } = await axios.post<{
-      coin: {
-        id: string;
-        amount: string;
-        assetId: string;
-        owner: string;
-        blockCreated: string;
-        txCreatedIdx: string;
-      };
-      jobId: string;
-      utxoId: string;
-    }>(`http://167.71.42.88:3000/allocate-coin`);
-    if (!data?.coin) {
-      throw new Error("No coin found");
-    }
 
-    if (!data.jobId) {
-      throw new Error("No jobId found");
-    }
-    const gasCoin: Coin = {
-      id: data.coin.id,
-      amount: bn(data.coin.amount),
-      assetId: data.coin.assetId,
-      owner: Address.fromAddressOrString(data.coin.owner),
-      blockCreated: bn(data.coin.blockCreated),
-      txCreatedIdx: bn(data.coin.txCreatedIdx),
-    };
-    console.log("gasCoin", gasCoin);
+    const paymaster = usePaymaster();
+    const { maxValuePerCoin } = await paymaster.metadata();
+    const { coin: gasCoin, jobId } = await paymaster.allocate();
+
     const amount = 10;
     const realAmount = amount / 1_000_000_000;
     const inputAmount = bn.parseUnits(realAmount.toFixed(9).toString());
@@ -148,26 +118,8 @@ export default function BuySeeds({
     );
     request.addChangeOutput(gasCoin.owner, provider.getBaseAssetId());
 
-    const response = await axios.post(`http://167.71.42.88:3000/sign`, {
-      request: request.toJSON(),
-      jobId: data.jobId,
-    });
-    if (response.status !== 200) {
-      throw new Error("Failed to sign transaction");
-    }
-    if (!response.data.signature) {
-      throw new Error("No signature found");
-    }
-    const gasInput = request.inputs.find((coin) => {
-      return coin.type === 0;
-    });
-    if (!gasInput) {
-      throw new Error("Gas coin not found");
-    }
-    console.log("gasInput", gasInput);
-    const wi = request.getCoinInputWitnessIndexByOwner(gasCoin.owner);
-    request.witnesses[wi as number] = response.data.signature;
-    console.log("request manually after coin", request.toJSON());
+    const {signature} = await paymaster.fetchSignature(request, jobId);
+    request.updateWitnessByOwner(gasCoin.owner, signature);
 
     const tx = await wallet.sendTransaction(request, {
       estimateTxDependencies: false,
