+++ frontend/src/components/modals/HarvestModal.tsx
@@ -59,42 +59,11 @@ export default function HarvestModal({
         try {
           // Try with gas station first
           const provider = await Provider.create(FUEL_PROVIDER_URL);
-          const { data: MetaDataResponse } = await axios.get<{
-            maxValuePerCoin: string;
-          }>(`http://167.71.42.88:3000/metadata`);
-          const { maxValuePerCoin } = MetaDataResponse;
-          console.log("maxValuePerCoin", maxValuePerCoin);
-          if (!maxValuePerCoin) {
-            throw new Error("No maxValuePerCoin found");
-          }
-          const { data } = await axios.post<{
-            coin: {
-              id: string;
-              amount: string;
-              assetId: string;
-              owner: string;
-              blockCreated: string;
-              txCreatedIdx: string;
-            };
-            jobId: string;
-            utxoId: string;
-          }>(`http://167.71.42.88:3000/allocate-coin`);
-          if (!data?.coin) {
-            throw new Error("No coin found");
-          }
 
-          if (!data.jobId) {
-            throw new Error("No jobId found");
-          }
-          const gasCoin: Coin = {
-            id: data.coin.id,
-            amount: bn(data.coin.amount),
-            assetId: data.coin.assetId,
-            owner: Address.fromAddressOrString(data.coin.owner),
-            blockCreated: bn(data.coin.blockCreated),
-            txCreatedIdx: bn(data.coin.txCreatedIdx),
-          };
-          console.log("gasCoin", gasCoin);
+          const paymaster = usePaymaster();
+          const { maxValuePerCoin } = await paymaster.metadata();
+          const { coin: gasCoin, jobId } = await paymaster.allocate();
+
           const addressIdentityInput = {
             Address: {
               bits: Address.fromAddressOrString(
@@ -122,23 +91,10 @@ export default function HarvestModal({
           request.gasLimit = gasUsed;
           request.maxFee = maxFee;
           console.log(`Harvest Cost gasLimit: ${gasUsed}, Maxfee: ${maxFee}`);
-          const response = await axios.post(`http://167.71.42.88:3000/sign`, {
-            request: request.toJSON(),
-            jobId: data.jobId,
-          });
-          if (response.status !== 200) {
-            throw new Error("Failed to sign transaction");
-          }
-          if (!response.data.signature) {
-            throw new Error("No signature found");
-          }
-          const gasInput = request.inputs.find((coin) => {
-            return coin.type === 0;
-          });
-          if (!gasInput) {
-            throw new Error("Gas coin not found");
-          }
-          request.witnesses[gasInput.witnessIndex] = response.data.signature;
+
+          const { signature } = await paymaster.fetchSignature(request, jobId);
+          request.updateWitnessByOwner(gasCoin.owner, signature);
+
           console.log("harvest request manually", request.toJSON());
           const tx = await wallet.sendTransaction(request);
           if (tx) {
