+++ frontend/src/components/modals/PlantModal.tsx
@@ -9,6 +9,7 @@ import Loading from "../Loading";
 import { Address, type Coin, Provider, bn } from "fuels";
 import { useWallet } from "@fuels/react";
 import axios from "axios";
+import { usePaymaster } from "../../hooks/usePaymaster";
 
 interface PlantModalProps {
   contract: FarmContract | null;
@@ -65,42 +66,11 @@ export default function PlantModal({
         try {
           // Try with gas station first
           const provider = await Provider.create(FUEL_PROVIDER_URL);
-          const { data: MetaDataResponse } = await axios.get<{
-            maxValuePerCoin: string;
-          }>(`http://167.71.42.88:3000/metadata`);
-          const { maxValuePerCoin } = MetaDataResponse;
-          console.log("maxValuePerCoin", maxValuePerCoin);
-          if (!maxValuePerCoin) {
-            throw new Error("No maxValuePerCoin found");
-          }
-          const { data } = await axios.post<{
-            coin: {
-              id: string;
-              amount: string;
-              assetId: string;
-              owner: string;
-              blockCreated: string;
-              txCreatedIdx: string;
-            };
-            jobId: string;
-            utxoId: string;
-          }>(`http://167.71.42.88:3000/allocate-coin`);
-          if (!data?.coin) {
-            throw new Error("No coin found");
-          }
 
-          if (!data.jobId) {
-            throw new Error("No jobId found");
-          }
-          const gasCoin: Coin = {
-            id: data.coin.id,
-            amount: bn(data.coin.amount),
-            assetId: data.coin.assetId,
-            owner: Address.fromAddressOrString(data.coin.owner),
-            blockCreated: bn(data.coin.blockCreated),
-            txCreatedIdx: bn(data.coin.txCreatedIdx),
-          };
-          console.log("gasCoin", gasCoin);
+          const paymaster = usePaymaster();
+          const { maxValuePerCoin } = await paymaster.metadata();
+          const { coin: gasCoin, jobId } = await paymaster.allocate();
+
           const addressIdentityInput = {
             Address: {
               bits: Address.fromAddressOrString(
@@ -129,26 +99,10 @@ export default function PlantModal({
           request.gasLimit = gasUsed;
           request.maxFee = maxFee;
           console.log(`Plant Cost gasLimit: ${gasUsed}, Maxfee: ${maxFee}`);
-          const response = await axios.post(`http://167.71.42.88:3000/sign`, {
-            request: request.toJSON(),
-            jobId: data.jobId,
-          });
-          if (response.status !== 200) {
-            throw new Error("Failed to sign transaction");
-          }
 
-          if (!response.data.signature) {
-            throw new Error("No signature found");
-          }
-          console.log("response.data", response.data);
-          const gasInput = request.inputs.find((coin) => {
-            return coin.type === 0;
-          });
-          if (!gasInput) {
-            throw new Error("Gas coin not found");
-          }
+          const { signature } = await paymaster.fetchSignature(request, jobId);
+          request.updateWitnessByOwner(gasCoin.owner, signature);
 
-          request.witnesses[gasInput.witnessIndex] = response.data.signature;
           const tx = await wallet.sendTransaction(request);
           if (tx) {
             console.log("tx", tx);
