+++ frontend/src/components/modals/SellItem.tsx
@@ -8,6 +8,7 @@ import type { FarmContract } from "../../sway-api/contracts";
 import Loading from "../Loading";
 import { useWallet } from "@fuels/react";
 import axios from "axios";
+import { usePaymaster } from "../../hooks/usePaymaster";
 
 interface SellItemProps {
   contract: FarmContract | null;
@@ -42,42 +43,10 @@ export default function SellItem({
             ).toB256(),
           },
         };
-        const { data: MetaDataResponse } = await axios.get<{
-          maxValuePerCoin: string;
-        }>(`http://167.71.42.88:3000/metadata`);
-        const { maxValuePerCoin } = MetaDataResponse;
-        console.log("maxValuePerCoin", maxValuePerCoin);
-        if (!maxValuePerCoin) {
-          throw new Error("No maxValuePerCoin found");
-        }
-        const { data } = await axios.post<{
-          coin: {
-            id: string;
-            amount: string;
-            assetId: string;
-            owner: string;
-            blockCreated: string;
-            txCreatedIdx: string;
-          };
-          jobId: string;
-          utxoId: string;
-        }>(`http://167.71.42.88:3000/allocate-coin`);
-        if (!data?.coin) {
-          throw new Error("No coin found");
-        }
 
-        if (!data.jobId) {
-          throw new Error("No jobId found");
-        }
-        const gasCoin: Coin = {
-          id: data.coin.id,
-          amount: bn(data.coin.amount),
-          assetId: data.coin.assetId,
-          owner: Address.fromAddressOrString(data.coin.owner),
-          blockCreated: bn(data.coin.blockCreated),
-          txCreatedIdx: bn(data.coin.txCreatedIdx),
-        };
-        console.log("gasCoin", gasCoin);
+        const paymaster = usePaymaster();
+        const { maxValuePerCoin } = await paymaster.metadata();
+        const { coin: gasCoin, jobId } = await paymaster.allocate();
 
         const scope = contract.functions.sell_item(
           seedType,
@@ -110,23 +79,10 @@ export default function SellItem({
           provider.getBaseAssetId()
         );
         request.addChangeOutput(gasCoin.owner, provider.getBaseAssetId());
-        const response = await axios.post(`http://167.71.42.88:3000/sign`, {
-          request: request.toJSON(),
-          jobId: data.jobId,
-        });
-        if (response.status !== 200) {
-          throw new Error("Failed to sign transaction");
-        }
-        if (!response.data.signature) {
-          throw new Error("No signature found");
-        }
-        const gasInput = request.inputs.find((coin) => {
-          return coin.type === 0;
-        });
-        if (!gasInput) {
-          throw new Error("Gas coin not found");
-        }
-        request.witnesses[gasInput.witnessIndex] = response.data.signature;
+
+        const { signature } = await paymaster.fetchSignature(request, jobId);
+        request.updateWitnessByOwner(gasCoin.owner, signature);
+
         const tx = await wallet.sendTransaction(request);
         console.log("tx", tx);
       } catch (err) {
